+++ react/components/admin/pages/Form/Form.tsx
@@ -95,6 +95,14 @@ const messages = defineMessages({
     defaultMessage: 'Keywords',
     id: 'admin/pages.admin.pages.form.field.meta.keywords',
   },
+  seoRobots: {
+    defaultMessage: 'Robots',
+    id: 'admin/store.metaTagRobots.title',
+  },
+  seoRobotsHelp: {
+    defaultMessage: 'Robots',
+    id: 'admin/store.metaTagRobots.description',
+  },
   bindingSelectorTitle: {
     defaultMessage: 'Binding',
     id: 'admin/pages.admin.pages.form.binding-selector.title',
@@ -217,9 +225,16 @@ const Form: React.FunctionComponent<Props> = ({
             placeholder=""
             value={data.metaTagKeywords}
           />
+          <FormFieldSeparator />
+          <Input
+            disabled={!!data.context}
+            label={intl.formatMessage(messages.seoRobots)}
+            onChange={detailChangeHandlerGetter('metaTagRobots')}
+            value={data.metaTagRobots}
+            helpText={intl.formatMessage(messages.seoRobotsHelp)}
+          />          
         </>
-      )}
-      <FormFieldSeparator />
+      )}      
       <SeparatorWithLine />
       <ConditionalTemplateSection
         intl={intl}