+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Add
+- Robots to be editable in page's form
+
+
 ## [4.52.0] - 2023-11-28
 
 ### Added
 
 - Translations for the `fetchSponsoredProductsOnSearch` advanced setting.
+
 
 ## [4.51.2] - 2023-07-10